requirements.txt
Flask==2.0.1
python-dotenv==0.19.1
spotipy==2.19.0
python-qobuz==0.1.2
requests==2.26.0
gunicorn==20.1.0